move useUser hook to hooks folder and added getUser function
